requirements.txt
openai
pyyaml
py-llm-core
